refactor - switch to sequelize
