+++ packages/server/src/common/Error.types.ts
