--- packages/server/src/entities.ts
@@ -1,1 +1,0 @@
-export * from "./modules/user/models/User"
