--- packages/server/src/migration/1613292096351-PostRefactoring.ts
@@ -1,11 +1,0 @@
-import {MigrationInterface, QueryRunner} from "typeorm";
-
-export class PostRefactoring1613292096351 implements MigrationInterface {
-
-    public async up(queryRunner: QueryRunner): Promise<void> {
-    }
-
-    public async down(queryRunner: QueryRunner): Promise<void> {
-    }
-
-}
