--- packages/server/src/migration/1614512799350-DateAdd.ts
@@ -1,22 +1,0 @@
-import {MigrationInterface, QueryRunner} from "typeorm";
-
-export class DateAdd1614512799350 implements MigrationInterface {
-    name = 'DateAdd1614512799350'
-
-    public async up(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "user" ADD "created" TIMESTAMP NOT NULL DEFAULT now()`);
-        await queryRunner.query(`ALTER TABLE "user" ADD "updated" TIMESTAMP NOT NULL DEFAULT now()`);
-        await queryRunner.query(`ALTER TABLE "activity_record" ADD "created" TIMESTAMP NOT NULL DEFAULT now()`);
-        await queryRunner.query(`ALTER TABLE "post" ADD "created" TIMESTAMP NOT NULL DEFAULT now()`);
-        await queryRunner.query(`ALTER TABLE "post" ADD "updated" TIMESTAMP NOT NULL DEFAULT now()`);
-    }
-
-    public async down(queryRunner: QueryRunner): Promise<void> {
-        await queryRunner.query(`ALTER TABLE "post" DROP COLUMN "updated"`);
-        await queryRunner.query(`ALTER TABLE "post" DROP COLUMN "created"`);
-        await queryRunner.query(`ALTER TABLE "activity_record" DROP COLUMN "created"`);
-        await queryRunner.query(`ALTER TABLE "user" DROP COLUMN "updated"`);
-        await queryRunner.query(`ALTER TABLE "user" DROP COLUMN "created"`);
-    }
-
-}
