--- packages/server/src/modules/activity/models/ActivityData.model.ts
@@ -1,18 +1,0 @@
-import { Field, ID, Int, ObjectType } from "type-graphql";
-import { BaseEntity, Column, Entity, PrimaryGeneratedColumn } from "typeorm";
-
-@Entity()
-@ObjectType()
-export class ActivityData extends BaseEntity {
-  @Field(() => ID)
-  @PrimaryGeneratedColumn("uuid")
-  id: number;
-
-  @Field(() => String)
-  @Column()
-  type: string;
-
-  @Field(() => Int)
-  @Column()
-  f_id: number;
-}
