+++ packages/server/src/server.ts
@@ -8,11 +8,10 @@ import { verify } from "jsonwebtoken";
 import * as path from "path";
 import "reflect-metadata";
 import { buildSchema } from "type-graphql";
-// import { createConnection } from "typeorm";
 import { GenerateAuthTokens, TokenPayload } from "./helpers/auth/auth_tokens";
 import { User } from "./models/User";
-import { customAuthChecker } from "./modules/common/authChecker";
-import { ContextType } from "./modules/common/types/Context.type";
+import { customAuthChecker } from "./common/authChecker";
+import { ContextType } from "./common/Context.type";
 import { ActivityResolver, PostResolver, UserResolver } from "./resolvers";
 import { sequelize } from "./sequelize";
 
@@ -33,7 +32,6 @@ const publicKEYREFRESH = readFileSync(
 
 const main = async () => {
 	// Create db connection
-	// await createConnection();
 	await sequelize.sync();
 
 	const app = express();
